deprecated/ml-service/requirements.txt
bilibili-api-python>=16.0.0
aiosqlite>=0.19.0
asyncpg>=0.28.0

# Discovery pipeline
httpx>=0.24.0
ollama>=0.4.0
pydantic>=2.0.0

# Cloud LLM backends (optional)
openai>=1.0.0
anthropic>=0.20.0

# ML Training
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
lightgbm>=4.1.0
joblib>=1.3.0

# PyTorch embedding fine-tuning + neural predictor
torch>=2.0.0
transformers>=4.30.0

# LoRA fine-tuning (optional — only needed for `fine-tune` command)
peft>=0.7.0
bitsandbytes>=0.41.0
trl>=0.7.0
datasets>=2.14.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
